requirements.txt
# Core dependencies
pymongo==4.6.0
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
PyYAML==6.0.1

# Scraping
apify-client==1.7.1

# AI/ML
google-generativeai==0.8.3
langdetect==1.0.9

# Date handling
python-dateutil==2.8.2

# Logging
colorlog==6.8.2